requirements.txt
asgiref==3.8.1
Django==4.2.21
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
qrcode==8.2
sqlparse==0.5.3
